finishing test

+++ testai/running-club/src/main/java/lt/techin/running/club/controller/EventController.java
@@ -1,21 +1,27 @@
 package lt.techin.running.club.controller;
 
-import lt.techin.running.club.dto.RunningEventRequestDTO;
-import lt.techin.running.club.dto.RunningEventResponseDTO;
-import lt.techin.running.club.dto.UserRequestDTO;
-import lt.techin.running.club.dto.UserResponseDTO;
+import lt.techin.running.club.dto.*;
 import lt.techin.running.club.dto.mapper.EventMapper;
+import lt.techin.running.club.dto.mapper.RegistrationMapper;
 import lt.techin.running.club.dto.mapper.UserMapper;
+import lt.techin.running.club.model.Registration;
 import lt.techin.running.club.model.Role;
 import lt.techin.running.club.model.RunningEvent;
 import lt.techin.running.club.model.User;
 import lt.techin.running.club.service.EventService;
+import lt.techin.running.club.service.RegistrationService;
+import lt.techin.running.club.service.UserService;
 import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @RestController
@@ -23,10 +29,14 @@ import java.util.List;
 public class EventController {
 
   private final EventService eventService;
+  private final UserService userService;
+  private final RegistrationService registrationService;
 
   @Autowired
-  public EventController(EventService eventService) {
+  public EventController(EventService eventService, UserService userService, RegistrationService registrationService) {
     this.eventService = eventService;
+    this.userService = userService;
+    this.registrationService = registrationService;
   }
 
   @GetMapping("/events")
@@ -52,8 +62,47 @@ public class EventController {
   }
 
   @DeleteMapping("/events/{id}")
-  public void deleteEvent(@PathVariable long id) {
+  public ResponseEntity deleteEvent(@PathVariable long id) {
+    if (!eventService.existsById(id)) {
+      return ResponseEntity.notFound().build();
+    }
     eventService.removeEvent(id);
+    return ResponseEntity.ok().build();
+  }
+
+  @PostMapping("/events/{id}/register")
+  public ResponseEntity<RegistrationRequestDTO> register(@PathVariable long id) {
+    User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+    System.out.println(user.getUsername());
+
+    Date newDate = new Date();
+    Registration newRegistration = new Registration(user, eventService.findEvent(id), new Timestamp(newDate.getTime()));
+
+    registrationService.addRegistration(newRegistration);
+
+    return ResponseEntity.created(ServletUriComponentsBuilder.fromCurrentRequest()
+                    .path("/{id}")
+                    .buildAndExpand(newRegistration.getId())
+                    .toUri())
+            .body(RegistrationMapper.toRegistrationRequestDTO(newRegistration));
+  }
+
+  @GetMapping("/events/{id}/participants")
+  public ResponseEntity<List<UserParticipantsDTO>> getEventParticipants(@PathVariable long id) {
+    List<User> users = userService.findUsers();
+    List<UserParticipantsDTO> participatingUsers = new ArrayList<UserParticipantsDTO>();
+
+    for (User user : users) {
+      for (Registration registration : user.getRegistrations()) {
+        if (registration.getRunningEvent().getId() == id) {
+          participatingUsers.add(new UserParticipantsDTO(user.getId(), user.getUsername()));
+          break;
+        }
+      }
+    }
+
+    return ResponseEntity.ok(participatingUsers);
   }
 
 }
